+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # frictionless 1.0.2
 
-- Add `testthat::skip_if_offline()` to selected tests and verbosely include 
+- Add `skip_if_offline()` to selected tests and verbosely include 
   output in vignette examples, to avoid CRAN errors caused by timeouts (#116).
 
 # frictionless 1.0.1